Moved ValidationHelper to source
